+++ src/test/java/org/roko/erp/controllers/CodeSerieControllerTest.java
@@ -23,9 +23,9 @@ import org.springframework.web.servlet.view.RedirectView;
 
 public class CodeSerieControllerTest {
     
-    private static final Long TEST_PAGE = 123l;
+    private static final int TEST_PAGE = 123;
 
-    private static final long TEST_COUNT = 345l;
+    private static final int TEST_COUNT = 345;
 
     private static final String TEST_CODE = "test-code";
     private static final String TEST_NAME = "test-name";
