+++ src/test/java/org/roko/erp/controllers/CustomerControllerTest.java
@@ -34,9 +34,8 @@ public class CustomerControllerTest {
 
     private static final String EXPECTED_CUSTOMER_LIST_TEMPLATE = "customerList.html";
 
-    private static final Long TEST_PAGE = 123l;
-
-    private static final Long TEST_CUSTOMER_COUNT = 234l;
+    private static final int TEST_PAGE = 123;
+    private static final int TEST_CUSTOMER_COUNT = 234;
 
     private static final String TEST_CODE = "test-code";
     private static final String TEST_NAME = "test-name";
