+++ src/test/java/org/roko/erp/controllers/PaymentMethodControllerTest.java
@@ -28,9 +28,8 @@ public class PaymentMethodControllerTest {
 
     private static final String CARD_TEMPLATE = "paymentMethodCard.html";
 
-    private static final Long TEST_PAGE = 123l;
-
-    private static final long TEST_COUNT = 234l;
+    private static final int TEST_PAGE = 123;
+    private static final int TEST_COUNT = 234;
 
     private static final String TEST_CODE = "test-code";
     private static final String TEST_NAME = "test-name";
